Move `MagicFolderService` to its own module.
